+++ examples/scripts/battery_parameterisation/bayesian_feature_fitting.py
@@ -1,9 +1,18 @@
 import numpy as np
 import pybamm
 from ep_bolfi.models.solversetup import spectral_mesh_pts_and_method
-from pybamm import CasadiSolver, Experiment, print_citations
 
 import pybop
+
+"""
+This example demonstrates how to use EP-BOLFI to parameterise a PyBaMM model using
+a "feature"-based cost function. We use the term "feature" to describe a parameter
+obtained from fitting either the data or a candidate solution to a simpler model.
+Every evaluation of a feature-based cost function runs its own optimisation (based
+on the simpler model) to identify the value of the feature. The aim is to minimise
+the "feature distance" to identify the parameter values which produce a candidate
+solution with a feature value as close as possible to that of the data.
+"""
 
 # Define model and parameter values
 model = pybamm.lithium_ion.SPMe()
@@ -13,16 +22,14 @@ original_D_p = parameter_values["Positive particle diffusivity [m2.s-1]"]
 
 # Set multivariate parameters (defined in model space)
 distribution = pybop.MultivariateLogNormal(
-    mean_log_x=[np.log(original_D_n), np.log(original_D_p)],
+    mean_log_x=[np.log(0.9 * original_D_n), np.log(1.1 * original_D_p)],
     covariance_log_x=[[np.log(2), 0.0], [0.0, np.log(2)]],
 )
 parameter_values["Negative particle diffusivity [m2.s-1]"] = pybop.Parameter(
-    initial_value=0.9 * original_D_n,
     transformation=pybop.LogTransformation(),
     distribution=pybop.MarginalDistribution(distribution, 0),
 )
 parameter_values["Positive particle diffusivity [m2.s-1]"] = pybop.Parameter(
-    initial_value=1.1 * original_D_p,
     transformation=pybop.LogTransformation(),
     distribution=pybop.MarginalDistribution(distribution, 1),
 )
@@ -32,13 +39,13 @@ submesh_types, var_pts, spatial_methods = spectral_mesh_pts_and_method(10, 10, 1
 simulator = pybop.pybamm.Simulator(
     model=model,
     parameter_values=parameter_values,
-    protocol=Experiment(
+    protocol=pybamm.Experiment(
         [
             "Discharge at 1.0 C for 15 minutes (1 second period)",
             "Rest for 15 minutes (1 second period)",
         ]
     ),
-    solver=CasadiSolver(
+    solver=pybamm.CasadiSolver(
         rtol=1e-5,
         atol=1e-5,
         root_tol=1e-3,
@@ -65,15 +72,15 @@ synthetic_data = simulator.solve(
 dataset = pybop.import_pybamm_solution(synthetic_data)
 
 ICI_cost = pybop.SquareRootFeatureDistance(
-    dataset["Time [s]"],
-    dataset["Voltage [V]"],
+    dataset=dataset,
+    target="Voltage [V]",
     feature="inverse_slope",
     time_start=0,
     time_end=90,
 )
 GITT_cost = pybop.SquareRootFeatureDistance(
-    dataset["Time [s]"],
-    dataset["Voltage [V]"],
+    dataset=dataset,
+    target="Voltage [V]",
     feature="inverse_slope",
     time_start=901,
     time_end=991,
@@ -89,9 +96,9 @@ if __name__ == "__main__":
     options = pybop.EPBOLFIOptions(
         ep_iterations=2,
         ep_total_dampening=0,
-        bolfi_initial_sobol_samples=10,
-        bolfi_optimally_acquired_samples=10,
-        bolfi_posterior_effective_sample_size=10,
+        bolfi_initial_sobol_samples=8,
+        bolfi_optimally_acquired_samples=8,
+        bolfi_posterior_effective_sample_size=8,
         posterior_gelman_rubin_threshold=1.2,
         verbose=True,
         model_parameter_boundaries={
@@ -107,8 +114,17 @@ if __name__ == "__main__":
     )
     optim = pybop.EP_BOLFI(problem, options=options)
     result = optim.run()
+    print("True values:", [original_D_n, original_D_p])
 
-    pybop.plot.convergence(result, yaxis={"type": "log"})
-    pybop.plot.parameters(result, yaxis={"type": "log"}, yaxis2={"type": "log"})
+    # Plot the optimisation result
+    result.plot_convergence(yaxis={"type": "log"})
+    result.plot_parameters(yaxis={"type": "log"}, yaxis2={"type": "log"})
 
-    print_citations()
+    # Plot the prior and posterior distributions
+    pybop.plot.distribution(result.problem.parameters, result.posterior)
+
+    # Plot predictions for a set of inputs sampled from the posterior
+    fig = result.plot_predictive(show=False)
+    fig[0].show()
+
+    pybamm.print_citations()