+++ pybop/plot/__init__.py
@@ -11,3 +11,5 @@ from .problem import problem
 from .nyquist import nyquist
 from .voronoi import surface
 from .samples import trace, chains, posterior, summary_table
+from .predictive import predictive
+from .distribution import distribution