+++ pybop/plot/plotly_manager.py
@@ -33,6 +33,7 @@ class PlotlyManager:
         """
         self.go = None
         self.pio = None
+        self.px = None
         self.make_subplots = None
         self.ensure_plotly_installed()
         self.check_renderer_settings()
@@ -43,12 +44,14 @@ class PlotlyManager:
         Check if Plotly is installed and import necessary modules; prompt for installation if missing.
         """
         try:
+            import plotly.express as px
             import plotly.graph_objs as go
             import plotly.io as pio
             from plotly.subplots import make_subplots
 
             self.go = go
             self.pio = pio
+            self.px = px
             self.make_subplots = make_subplots
         except ImportError:
             self.prompt_for_plotly_installation()
@@ -86,12 +89,14 @@ class PlotlyManager:
         """
         Import Plotly modules and set the default renderer after installation.
         """
+        import plotly.express as px
         import plotly.graph_objs as go
         import plotly.io as pio
         from plotly.subplots import make_subplots
 
         self.go = go
         self.pio = pio
+        self.px = px
         self.make_subplots = make_subplots
         if pio.renderers.default == "":
             pio.renderers.default = "browser"