+++ tests/integration/test_model_experiment_changes.py
@@ -1,7 +1,6 @@
 import numpy as np
 import pybamm
 import pytest
-from scipy import stats
 
 import pybop
 
@@ -19,9 +18,7 @@ class TestModelAndExperimentChanges:
                 {
                     "Negative particle radius [m]": pybop.Parameter(  # geometric parameter
                         distribution=pybop.Gaussian(
-                            6e-06,
-                            0.1e-6,
-                            truncated_at=[1e-6, 9e-6],
+                            6e-06, 0.1e-6, truncated_at=[1e-6, 9e-6]
                         ),
                         initial_value=5.86e-6,
                     ),
@@ -32,9 +29,7 @@ class TestModelAndExperimentChanges:
                 {
                     "Positive particle diffusivity [m2.s-1]": pybop.Parameter(  # non-geometric parameter
                         distribution=pybop.Gaussian(
-                            3.43e-15,
-                            1e-15,
-                            truncated_at=[1e-15, 5e-15],
+                            3.43e-15, 1e-15, truncated_at=[1e-15, 5e-15]
                         ),
                         initial_value=4e-15,
                     ),
@@ -158,7 +153,7 @@ class TestModelAndExperimentChanges:
         parameter_values.update(
             {
                 "Negative electrode active material volume fraction": pybop.Parameter(
-                    distribution=stats.norm(loc=0.68, scale=0.05),
+                    distribution=pybop.Gaussian(0.68, 0.05)
                 )
             }
         )
@@ -181,7 +176,7 @@ class TestModelAndExperimentChanges:
         parameter_values.update(
             {
                 "Negative electrode active material volume fraction": pybop.Parameter(
-                    distribution=stats.norm(loc=0.68, scale=0.05),
+                    distribution=pybop.Gaussian(0.68, 0.05)
                 )
             }
         )